core code updates - minor enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@5552 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/RangeVariable.java
@@ -1135,7 +1135,7 @@ public class RangeVariable {
 
             return currentRow == null ? 0
                                       : ((long) rangeVar.rangeTable.getId() << 32)
-                                        + ((long) currentRow.getPos());
+                                        + currentRow.getPos();
         }
 
         public TableBase getCurrentTable() {
